+++ exposed-jdbc/src/main/kotlin/org/jetbrains/exposed/sql/statements/jdbc/JdbcDatabaseMetadataImpl.kt
@@ -57,7 +57,7 @@ class JdbcDatabaseMetadataImpl(database: String, val metadata: DatabaseMetaData)
     override val supportsSelectForUpdate: Boolean by lazyMetadata { supportsSelectForUpdate() }
 
     override val identifierManager: IdentifierManagerApi by lazyMetadata {
-        identityManagerCache.computeIfAbsent(url) {
+        identityManagerCache.getOrPut(url) {
             JdbcIdentifierManager(this)
         }
     }
@@ -171,6 +171,8 @@ class JdbcDatabaseMetadataImpl(database: String, val metadata: DatabaseMetaData)
 
     private fun sanitizedDefault(defaultValue: String) = when (currentDialect) {
         is SQLServerDialect -> defaultValue.trim('(', ')', '\'')
+        is OracleDialect -> defaultValue.trim().trim('\'')
+        is MysqlDialect -> defaultValue.substringAfter("b'").trim('\'').trim()
         else -> defaultValue.trim('\'').trim()
     }
 
@@ -225,7 +227,9 @@ class JdbcDatabaseMetadataImpl(database: String, val metadata: DatabaseMetaData)
                 } ?: return@iterate null // Do not crash if there are missing fields in Exposed's tables
                 val constraintName = getString("FK_NAME")!!
                 val targetTableName = getString("PKTABLE_NAME")!!
-                val targetColumnName = identifierManager.quoteIdentifierWhenWrongCaseOrNecessary(getString("PKCOLUMN_NAME")!!)
+                val targetColumnName = identifierManager.quoteIdentifierWhenWrongCaseOrNecessary(
+                    identifierManager.inProperCase(getString("PKCOLUMN_NAME")!!)
+                )
                 val targetColumn = allTables.getValue(targetTableName).columns.first {
                     identifierManager.quoteIdentifierWhenWrongCaseOrNecessary(it.nameInDatabaseCase()) == targetColumnName
                 }
@@ -238,7 +242,11 @@ class JdbcDatabaseMetadataImpl(database: String, val metadata: DatabaseMetaData)
                     onDelete = constraintDeleteRule,
                     name = constraintName
                 )
-            }.filterNotNull()
+            }
+                .filterNotNull()
+                .groupBy { it.fkName }
+                .values
+                .map { it.reduce(ForeignKeyConstraint::plus) }
         }
     }