+++ exposed-tests/src/test/kotlin/org/jetbrains/exposed/sql/tests/shared/ddl/CreateMissingTablesAndColumnsTests.kt
@@ -8,6 +8,8 @@ import org.jetbrains.exposed.sql.tests.DatabaseTestsBase
 import org.jetbrains.exposed.sql.tests.TestDB
 import org.jetbrains.exposed.sql.tests.currentDialectTest
 import org.jetbrains.exposed.sql.tests.inProperCase
+import org.jetbrains.exposed.sql.tests.shared.assertEqualCollections
+import org.jetbrains.exposed.sql.tests.shared.assertEqualLists
 import org.jetbrains.exposed.sql.tests.shared.assertEquals
 import org.jetbrains.exposed.sql.tests.shared.assertFailAndRollback
 import org.jetbrains.exposed.sql.tests.shared.assertFalse
@@ -126,6 +128,95 @@ class CreateMissingTablesAndColumnsTests : DatabaseTestsBase() {
                 it[id] = 3
                 it[foo] = "ABC"
             }
+
+            SchemaUtils.drop(t1)
+        }
+    }
+
+    @Test
+    fun testAddMissingColumnsStatementsChangeCasing() {
+        val t1 = object : Table("foo") {
+            val id = integer("idCol")
+
+            override val primaryKey = PrimaryKey(id)
+        }
+
+        val t2 = object : Table("foo") {
+            val id = integer("idcol")
+
+            override val primaryKey = PrimaryKey(id)
+        }
+
+        withDb {
+            if (db.supportsAlterTableWithAddColumn) {
+                SchemaUtils.createMissingTablesAndColumns(t1)
+
+                val missingStatements = SchemaUtils.addMissingColumnsStatements(t2)
+
+                val alterColumnWord = when (currentDialectTest) {
+                    is MysqlDialect -> "MODIFY COLUMN"
+                    is OracleDialect -> "MODIFY"
+                    else -> "ALTER COLUMN"
+                }
+
+                val expected = if (t1.id.nameInDatabaseCase() != t2.id.nameInDatabaseCase()) {
+                    "ALTER TABLE ${t2.nameInDatabaseCase()} $alterColumnWord ${t2.id.nameInDatabaseCase()} INT"
+                } else null
+
+                assertEquals(expected, missingStatements.firstOrNull())
+
+                SchemaUtils.drop(t1)
+            }
+        }
+    }
+
+    @Test
+    fun testAddMissingColumnsStatementsIdentical() {
+        val t1 = object : Table("foo") {
+            val id = integer("idcol")
+
+            override val primaryKey = PrimaryKey(id)
+        }
+
+        val t2 = object : Table("foo") {
+            val id = integer("idcol")
+
+            override val primaryKey = PrimaryKey(id)
+        }
+
+        withDb {
+//        withDb(db = listOf(TestDB.H2)) {
+            SchemaUtils.createMissingTablesAndColumns(t1)
+
+            val missingStatements = SchemaUtils.addMissingColumnsStatements(t2)
+
+            assertEqualCollections(missingStatements, emptyList())
+
+            SchemaUtils.drop(t1)
+        }
+    }
+
+    @Test
+    fun testAddMissingColumnsStatementsIdentical2() {
+        val t1 = object : Table("foo") {
+            val id = integer("idCol")
+
+            override val primaryKey = PrimaryKey(id)
+        }
+
+        val t2 = object : Table("foo") {
+            val id = integer("idCol")
+
+            override val primaryKey = PrimaryKey(id)
+        }
+
+        withDb {
+//        withDb(db = listOf(TestDB.H2)) {
+            SchemaUtils.createMissingTablesAndColumns(t1)
+
+            val missingStatements = SchemaUtils.addMissingColumnsStatements(t2)
+
+            assertEqualCollections(missingStatements, emptyList())
 
             SchemaUtils.drop(t1)
         }
@@ -265,6 +356,31 @@ class CreateMissingTablesAndColumnsTests : DatabaseTestsBase() {
         }
     }
 
+    private enum class TestEnum { A, B, C }
+
+    @Test
+    fun `check that running addMissingTablesAndColumns multiple time doesnt affect schema`() {
+        val table = object : Table("defaults2") {
+            val bool1 = bool("boolCol1").default(false)
+            val bool2 = bool("boolCol2").default(true)
+            val int = integer("intCol").default(12345)
+            val float = float("floatCol").default(123.45f)
+            val decimal = decimal("decimalCol", 10, 1).default(BigDecimal.TEN)
+            val string = varchar("varcharCol", 50).default("12345")
+            val enum1 = enumeration("enumCol1", TestEnum::class).default(TestEnum.B)
+            val enum2 = enumerationByName("enumCol2", 25, TestEnum::class).default(TestEnum.B)
+        }
+
+        withDb {
+            try {
+                SchemaUtils.create(table)
+                assertEqualLists(emptyList(), SchemaUtils.statementsRequiredToActualizeScheme(table))
+            } finally {
+                SchemaUtils.drop(table)
+            }
+        }
+    }
+
     @Test
     fun createTableWithMultipleIndexes() {
         withDb {
@@ -291,6 +407,23 @@ class CreateMissingTablesAndColumnsTests : DatabaseTestsBase() {
         }
     }
 
+    @Test
+    fun testCamelCaseForeignKeyCreation() {
+        val ordersTable = object : IntIdTable("tmporders") {
+            val traceNumber = char("traceNumber", 10).uniqueIndex()
+        }
+        val receiptsTable = object : IntIdTable("receipts") {
+            val traceNumber = reference("traceNumber", ordersTable.traceNumber)
+        }
+
+        withDb {
+            SchemaUtils.createMissingTablesAndColumns(ordersTable, receiptsTable)
+            assertTrue(ordersTable.exists())
+            assertTrue(receiptsTable.exists())
+            SchemaUtils.drop(ordersTable, receiptsTable)
+        }
+    }
+
     object MultipleIndexesTable : Table("H2_MULTIPLE_INDEXES") {
         val value1 = varchar("value1", 255)
         val value2 = varchar("value2", 255)
@@ -301,7 +434,7 @@ class CreateMissingTablesAndColumnsTests : DatabaseTestsBase() {
         }
     }
 
-    @Test fun testCreateTableWithReferenceMutipleTimes() {
+    @Test fun testCreateTableWithReferenceMultipleTimes() {
         withTables(PlayerTable, SessionsTable) {
             SchemaUtils.createMissingTablesAndColumns(PlayerTable, SessionsTable)
             SchemaUtils.createMissingTablesAndColumns(PlayerTable, SessionsTable)
@@ -362,6 +495,29 @@ class CreateMissingTablesAndColumnsTests : DatabaseTestsBase() {
             } finally {
                 SchemaUtils.drop(usersTable)
             }
+        }
+    }
+
+    object CompositePrimaryKeyTable : Table("H2_COMPOSITE_PRIMARY_KEY") {
+        val idA = varchar("id_a", 255)
+        val idB = varchar("id_b", 255)
+        override val primaryKey = PrimaryKey(idA, idB)
+    }
+
+    object CompositeForeignKeyTable : Table("H2_COMPOSITE_FOREIGN_KEY") {
+        val idA = varchar("id_a", 255)
+        val idB = varchar("id_b", 255)
+
+        init {
+            foreignKey(idA, idB, target = CompositePrimaryKeyTable.primaryKey)
+        }
+    }
+
+    @Test
+    fun testCreateCompositePrimaryKeyTableAndCompositeForeignKeyTableMultipleTimes() {
+        withTables(CompositePrimaryKeyTable, CompositeForeignKeyTable) {
+            SchemaUtils.createMissingTablesAndColumns(CompositePrimaryKeyTable, CompositeForeignKeyTable)
+            SchemaUtils.createMissingTablesAndColumns(CompositePrimaryKeyTable, CompositeForeignKeyTable)
         }
     }
 }